+++ bin/fake_data.py
@@ -7,6 +7,7 @@ import sys
 
 test = load_proto("probe_testxyz_100_proto")
 test2 = load_proto("probe_testimage_100_proto")
+ARRAY_TEST = load_proto("image_mood_classification_100_proto")
 
 HOST = "localhost:5006" #default
 if len(sys.argv) > 1:
@@ -33,6 +34,23 @@ while True:
             sleep(0.1)
         except requests.exceptions.ConnectionError: #allow this script to keep running when developing and shutting on/off the server
             pass
+
+
+    #try a test with arrays
+    new_dict = {}
+    for listKey in ['good', 'bad', 'ugly']:
+        new_dict[listKey] = random.random()
+    msg1 = ARRAY_TEST.ImageTagSet(image = [0,0,0], tag=list(new_dict.keys()), score=list(new_dict.values()))
+    msg1b = msg1.SerializeToString()
+    try:
+        r = requests.post(url,
+                          data=msg1b,
+                          headers={"PROTO-URL": "http://cognita-nexus01.eastus.cloudapp.azure.com:8081/repository/repo_cognita_model_raw/image-mood-classification/1.0.0/image-mood-classification-1.0.0-proto",
+                                   "Message-Name": "ImageTagSet"})
+        print(r.status_code)
+        assert(msg1b == r.content)
+    except requests.exceptions.ConnectionError: #allow this script to keep running when developing and shutting on/off the server
+        pass
 
     msg2 = test2.TransformedImagePNG()
     theimage = random.choice(["/tmp/1.png","/tmp/2.png","/tmp/3.png","/tmp/4.png","/tmp/5.png","/tmp/6.png","/tmp/7.png","/tmp/8.png","/tmp/9.png","/tmp/10.png"])