+++ docs/release-notes.rst
@@ -27,6 +27,17 @@ The format is based on `Keep a Changelog <http://keepachangelog.com/>`__
 and this project adheres to `Semantic
 Versioning <http://semver.org/>`__.
 
+[1.4.0] - 3/25/18
+-------
+
+- Handle "repeated" keyword
+- Better raw display
+
+[1.3.1] - 3/24/18
+-------
+
+- Swap out marshal for pickle
+
 [1.3.0]
 -------